New settings system
You can now click on your name to edit it and on the little badge to change the type of alien you play.
Also, integrates the font Open Sans in the repo, so for a developper it is loaded from the file sytem instead of from Google, which helps when your Internet connection is crappy.