Fix bugs, add SVG optimization script and update to latest vinage
Fixes fuel bar not going down and replace it with a <meter> element.
Fixes jetpack not being heard.

Update vinage to the latest version. The wrapped universe feature is not
used though because it is still buggy.

optimize_svg.sh will be used to batch optimize SVGs. It needs scour in
order to run.