Improve protocol
Now players must register their name to the server with SET_NAME before connecting to a lobby.
This message can also be used when already connected to a lobby; in this case the server broadcasts SET_NAME_BROADCAST to every connected player so the name of the player can be updated locally.
Because of these changes, CONNECT has been simplified and PLAYERS_SETTINGS removed.