Make the warping at the edges of the world invisible
The transition isn't smooth tough. This has to do with the prediction.

Note that the enemies and the players do not appear anymore on Chrome.
This happens actually happens in Vinage because Chrome doesn't support the ES6 proxies yet. Since the next version which will be released in two weeks brings them, no workarounds will be made.