Complete the transition to the binary protocol
JumpSuit doesn't use JSON messages anymore!
This means that the game will be way less laggy!

The protocol needs some refinement though, because some stuff it sends is useless and because using a float32 to send angle data is not the best we can do.
The implementation needs work too because its "API" has been made to look somewhat like what was done in JSON to facilitate the transition, but that is useless now.