Revert previous commit but reuse code
Actually, having to resolve before connecting was a bad idea because it delays the connection to a game server (which is already delayed because the socket can take some time before opening).
And the amount of byte saved is ludicrous compared to what could be saved by other means (minimification, re-encoding of audio files, etc).