Use opus files instead of vorbis
Opus is incredible, from my tests an opus file half the size of a vorbis file has about the same quality. But the vorbis files used also were way too high quality, so it has been possible to reduce their sizes by a factor of 8.
Chrome only supports opus on <audio> elements (https://bugs.chromium.org/p/chromium/issues/detail?id=482934) so instead of loading files with an XHR it's done with an audio element, which is a bit hacky but hopefully works.