Fixes
Fix settings and info buttons not working when playing.
Remove hack on audio.js to make opus files work on Chrome. It worked okay on the desktop but on mobile it made the game crash because it used play() on a HTML element despite that "play() can only be initiated by a user gesture". So now OGG files are played on Chrome (which is a shame because they are much bigger for an equivalent quality).
Fix a letter being entered in the chat when pressing the key to open it.
Changes the hearts because they had too much details, which looked bad at the low resolution they are displayed.