Two sounds added, shot position update removed
I added two new sounds that are licensed under CC-Zero and need to be converted into .opus. Furthermore shots won't be longer updated by server, the server will just confirm a shots deletion and a shots spawning. Currently the server sends it with ADD_ENTITY but to make it synchroneous with the GAME_STATE it should be sent within the GAME_STATE...