Modifications to allow HTTP/2 use
This changes protocol relative URL to https://jumpsuit.space/something so that the index is still served as HTTP but resources are served as HTTPS.
This commit also add an option to use protocol relative URLs, which is easier for developpement.
More about this can be found in the footnotes of readme.md.