Layout and design improvements
I worked on style.css and layout.css in order to make the GUI look more responsive. When the window is scaled down, the menu-box and settings-box change their layouts. Also the minimap canvas issue on smaller screen is fixed too.
Furthermore the lobby search now highlights the matches with red and underlines them.
TODO: There is an issue with player names when connecting, where the last letter is cut off. Also submitting the current name again adds a homographId eventhough it's not necessary.