Game server automatically try to reconnect
In case the master server shuts down the game server tries to reconnect to the master server set in the config file every 5 seconds. Also if the server is registered at the master server it sends a message back to the game server. Further the master server sends ping's to the game server in order to get the latency between to the master server. This can be used to display the connection speed of the server.
Moreover I added an automatic connection close if the server doesn't send a message for 7 seconds. The client closes the connection stream and stops the game so that the menu is displayed again