Optimized MESSAGE.SCORES and added graphic options
I added checkboxes into the settings-box with whom the user can enable and disable particles and meteors in order to give some performance boost. Also changes on the weaponry settings are now saved in the localStorage and will be loaded whenever the page is loaded as well as the changes on the graphics settings.