Have the resources as a global
No no no, global variables aren't always bad :-)
In this case, since resources are loaded asynchrounously on the client not having `resources` as a global variable used to add complexity to module loading.
Which had to be dealt with on the server too, even though it loads resources synchronously!

The only thing to be careful with is that all code depending on `resources` must not run before they are loaded but it's easy to do.