Replace old mod loader with new one that works with ES6 modules
Because ES6 modules's path have to be static with Rollup (which makes sense), modules can no longer be loaded dynamically by the game server (which is a bit sad to be honest, but maybe once node supports ES6 modules it will be no longer be the case? I certainly hope so!).
So the mod loading has to be made at bundling-time. Basically the source is preprocessed to have the correct path to the selected mod set.