More module-related fixes
I realized a lot of this were set as globals in the ESLint config, which means a lot of objects were not reported to be unknown by the module.
Having removed these globals from the config I fixed some things but there is a lot of work to be done, especially on how to organize this plate of spaghetti. Maybe using a proven architectural pattern?