Fix bugs when leaving lobby + remove HUD management code
The same canvas was used for showing resource loading progress and the game itself, so the HUD had to be hidden/shown.
Since now loading progress is shown with HTML elements, in a div that sits on top of everything else, all the code that manages the HUD is useless and has been removed.