Update spec, taking how callbacks are used into account + implement Planet serializator
I want to make the serialization more modular, by splitting code into subpayloads serializators, but this also has to play nice with the callback API.

After running a few benchmark, I was able to confirm that the callback approach (ie: handing the deserialized values through callbacks) is faster than returning objects. Probably because there is less GC?
It is especially faster when what we are deserializing is a collection of values (ex: planets) because it means we don't have to go trough the collection twice: once for creating it and once for using the values in the client.