Updated message protocoll and player timeout
Since we're using some kind of UDP now to transmit data the way we sent the player controls was inappropriate so it had to be changed. I realized that the aimAngle can be merged into the playerControls soboth controls and the mouse angle are sent simultaneously.
Furthermore player didn't disconnect when they just closed their browser window and - I don't know why - the DataChannel kept being in `open` state eventhough there was no client anymore, obviously. So now the server looks for the activity and kills/removes everyone that hasn't sent data for 7 seconds (just like Timeouts work).